Add search query parser
This adds a search query parser using the Parslet Ruby library. This
allows us to use a formal grammar to define the makeup of search
queries.

The implemented parser defines rules for term and phrase based
queries.

For example, a query of 'flood map for planning' would be parsed as a
term based query. A query of '"flood map" for planning' would be parsed
a phrase of "flood map" and the terms "for planning".

See http://kschiess.github.io/parslet/