Handle all exceptions with the configured exceptions_app
Removes custom error handling for datasets and legacy datafiles. Rails
takes care of this for us.

The error handling is also far too aggresive. We're swallowing any
form of exception and communicating this downstream as not found which
is missleading. We should only say something is not found when we know
that is the case.