Renders summaries as markdown...
.. because they're markdown.

Currently all of the summaries are markdown, but ignore that and end up
squashed into a single paragraph/wall-of-text.

This PR renders the markdown to the page which means it will autolink,
and keep the appropriate spacing.

We can turn off certain features, so we can stop people using some
markdown features.