Order datafiles in datables by created_at desc
Datafiles for a dataset are stored in an inline array, but may be
displayed in groups (timeseries) or ungrouped (non-timeseries). Both
views use the datatable partial, to which I've added a sorting method on
the input datafiles array, using the lexicographic property of ISO8601
timestamps like created_at to sort and then reverse (most recent first).