Add additional information
Ensures correct licence information is displayed.

Works with the licence information we currently have in the index,
and also the information added to the index as a result of
https://github.com/alphagov/datagovuk_publish/pull/558.

Once the index has been rebuilt we can remove most of this logic and
use the information directly from the index.

https://trello.com/c/EwW9h6EU