Replace displayed_date helper with public_updated_at
We have made `public_updated_at` available for each Dataset, which does the job
of determining whether a Dataset has datafiles or not and selecting the right
public-facing timestamp to display. This means we no longer need the
`displayed_date` helper to figure this out, so we replace references to it in
the code with the `public_updated_at` value for a Dataset.