Refactor zendesk implementation with tests
The previous implementation used a set of client classes (models?!),
assigned to a global constant, which introduced an unnecessary amount
of indirection. Instead, we now have a Zendesk module that can return a
real or dummy client as appropriate.

Note that Zendesk will now be enabled for all non-dev/test environments,
as we should allow for manual testing in these environments.