Initialize all model attributes manually
Using ActiveModel to initialize model attributes would create issues
when a new attribute appeared in ES but was not defined in the model.
There are two approaches to avoid this: allow all attributes using
something like OpenStruct, or unpack the attributes manually in the
initializer (or whatever calls the initializer).

The approach taken here is to manually populate the model attributes,
which has the advantage of visibility and allows for extra mapping to be
done when the model is loaded.