Replace licence/licence_code with explicit fields
Licence information is now available in 4 explicit fields: licence_code,
licence_title, licence_url and licence_custom. The approach taken on the
show page is to try and show as much licence information as possible,
even if some of the fields were not completed.

The licence information is also now used in a nested CreativeWork
document as part of the linked data for the show page, although this
isn't tested as it's unclear that this feature is in use. Otherwise, the
CCZero test has been replaced in favour of testing different behaviours.