Increase the RAM available to the application
Running `cf events find-data-beta | grep find-data-beta | wc -l` shows
that the find application has crashed 26 times in the last 2 hours. Each
crash is logged as `Exited with status 137 (out of memory)`.

This PR increases the amount of memory available to each instance to
reduce the number of crashes. Unfortunately if an application crashes
during a request, the inflight request will fail, rather than it
being re-routed to a live instance.  So this means users (or systems)
are seeing errors of some form or other.