Allow map preview to load content images from any URL
Previously we tried to set a strong CSP so that map preview images
could only be loaded from OS and the WMS host. This is proving to be
too strict as we have WMS sources that link to other URIs for their
map data. For example
https://data.gov.uk/data/map-preview?e=1.8&n=55.8&s=50.0&url=http%3A%2F%2Fenvironment.data.gov.uk%2Fds%2Fwms%3FSERVICE%3DWMS%26INTERFACE%3DENVIRONMENT--6f59c1ce-cc11-43aa-b11d-e1c3ab43a192%26request%3DGetCapabilities&w=-5.7

This allows images to be loaded from any URI so that we can support
map previews.