Fix failing master tests due to chrome driver bug
This is a temporary fix to failing master tests as per
https://stackoverflow.com/questions/50642308/org-openqa-selenium-webdriverexception-unknown-error-devtoolsactiveport-file-d
as we have seen similar error messages on a selection
of tests:

```Selenium::WebDriver::Error::UnknownError:
unknown error: DevToolsActivePort file doesn't exist```

It looks like this was a recent issue introduced with an
upgrade of Chrome driver that conflicts with Selenium.

A fix has been put in to Selenium master in
https://github.com/SeleniumHQ/selenium/issues/6071
but has not yet been released as a new version. When a newer
version of Selenium has been released we can upgrade to it.