Update WMS requests to version 1.3.0
A number of GetFeatureInfo requests were failing on specific WMS
endpoints. It's thought that version 1.1.1 API calls were no longer
supported though we don't have visibility of these servers and have
requested more information from the environment agency.
Using WMS 1.3.0 via OpenLayers appears to fix this problem, some working
calls have been tested along with the failing ones and both seem to
work.
This commit sets the OpenLayers WMS version to 1.3.0, it also reinstates
setting the exception parameter as this is required by v1.3.0.
There was also an issue with infoFormat and exceptionFormat being set to
a global scope, this has been rectified as part of the commit.