Use API key for submitting Zendesk tickets
The current ZenDesk config was using both a password and an API key.
When both are submitted, it appears it default to using the password.

This password can expire, needing manual intervention to update.

By removing the password from the config, the API key will be used
for authentication instead.

Co-authored-by: Rosa Fox <rosa.fox@digital.cabinet-office.gov.uk>