Retrieve S3 bucket data for correct environment
Currently, we always fetch organograms data from the production S3
bucket regardless of the environment that we make the request from. This
causes staging to try to fetch production data and the request fails due
to CORS rules that don't allow this.

This commit fixes this issue by dynamically constructing the S3 bucket
name for the correct environment.

Trello card: https://trello.com/c/Vx11ucIl/1043-fix-display-of-organogram-visualisations-on-stagingdatagovuk

Co-Authored-By: Murilo Dal Ri <murilo.dalri@digital.cabinet-office.gov.uk>