Specify Accept-Encoding when calling URI().read
Without this, something seems to go wrong. I'm not sure what value is
being sent for this header (if any) without this method argument, nor
do I know if this is specific to some server configurations, but at
least with one [1] URL, there's a Zlib::BufError raised from inside of
open-uri if the Accept-Encoding parameter isn't specified.

I found a reference to Ruby supporting gzip and deflate, so I went
with accepting those.

1: https://emaps.eastleigh.gov.uk/inspire/wms.exe?_dc=1570638796663&request=GetCapabilities&service=WMS