Raise as 404 error if legacy datafile cannot be found
We are currently not catching this error, which throws a 500
internal server error when a datafile can't be found for a legacy
dataset. This is misleading to users and should be treated as a 404
error.

We have a number of datasets with this issue, which might affect
how Google ranks data.gov.uk.