Fix Elasticsearch URI not found in VCAP_SERVICES
Rails 6 deprecated the use of string keys to access the hashes generated
by config_for, so a change was done to use symbols instead.

However, an incorrect change was also done to the way the keys of
VCAP_SERVICES are accessed, which must be done with strings, resulting
in the app not being able to find the Elasticsearch URI.

Revert keys to strings to restore correct behaviour.