Document working staging URL
data.gov.uk has two different staging URLs. Given the lack of
resources and the infrequency of releases, I'm inclined to
document the (working) staging URL and to write a disclaimer
about the (currently not working) staging URL. This is to
unblock deploying to Production, which is now several commits
behind what is on Staging.