Rename CKAN_REDIRECTION_URL to CKAN_DOMAIN
The CKAN_REDIRECTION_URL is no longer a full URL, it's just the domain
which CKAN is running on. The code that does the redirect adds the other
parts of the URL (the protocol and the path).

As well as this, I want to add the CKAN Domain to the content security
policy on some pages. It wouldn't make sense to add a redirection URL to
the CSP, so it feels important that this is named accurately.

I also found that config.ckan_redirection_url wasn't used anywhere, so
I've removed it.