Add basic search page
The default behaviour for this page is to show all results if a search
term has not been provided. This initial search page does not have
working filters or a working search box, and the results do not link
to individual dataset pages. Their functionality will be added in
future commits.

We still want OpenSearch to be our default search method, but we can
access the alternative search using Solr via /search/solr.