Add sort filters to search page
By default the results returned are "best match" and the sort option
in the filter is automatically selected.

When the results are sorted by "most recent" we want to sort using
the `public_updated_at` (the equivalent field in Solr is
`metadata_modified`) and ordered by desc. This is passed as the
`sort` parameter in the Solr query.