Display data files on dataset page
If there are data files available they are displayed in a table in
ascending order. By default a maximum of 5 rows are shown.

There are some instances where the first data file does not have
data in Solr's `created_at` field. In OS this date matches the
date the dataset was created. We can use the `metadata_created`
field in Solr to get this information, and pass it to the
`SolrDatafile` model. This will also allow us to sort the files
using the `created_at` field.