Add Solr query to get an organisation by slug
Solr does not always have contact information available for an
organisation which is shown on the dataset page.

We can get this information from CKAN using a custom index in Solr
(`site_id:dgu_organisations`), which contains all the data we need
for each organisation. We can query Solr by passing the index and
name (slug) of the organisation as filter query (`fq`) parameters.