Get organisation information using Solr query
An organisation's contact and FOI details aren't always available
in the `organization` hash for a dataset, which is shown on the
dataset page. To solve this, we can now query using the custom
index in Solr. We can then use the existing `Organisation` model
to process the information.

The `Organisation` model is still being used for the OS version of
search and uses different names for some of it's fields.