Display additional information on dataset page
This is additional metadata information available if the dataset
is an INSPIRE dataset. In Solr it's stored in the `extras` field
within the validated data dict. Some additional processing is
required to extract the required fields as each field is stored
in the format { key: field_name, value: field_value }.