Update Solr search query to accept user input
The Solr query has been modified to accept user input, and treated
as a search phrase to look for datasets using the `title` and `notes`
fields. We also want to avoid including our custom index in the
search which would return organisations rather than datasets.

This search query needs to be modified to also search by single terms
(split a phrase by whitespace) if the user input contains multiple
words. So for example, if "animal health" is entered, Solr should
search for "animal" and "health" as well as "animal health".