List all organisations in publisher filter
This is the default behaviour when a user visits the search page
without entering a search term or selected any filters.

We can get this list by making a Solr query for the `title` field
for each organisation from the custom organisations
index (`site_id:dgu_organisations`) which returns an array of
organisation names.