Only show filter publishers included in results
If a user has entered a search term, the publisher filter will
only include organisations of datasets that are included in the
search results.

We can get this information by including the use of faceting in the
Solr query. We can specify the `organization` field as a facet, and
only return any where the minimum count is 1.