List most popular formats in format filter
This is the default list shown when a user visits the search page
without entering a search term or selected any filters.

Similar to `topics`, this is a fixed list independent to the initial
search results shown so we do not need to query it from Solr.

We can filter by a selected format by updating the Solr query so
that it searches for the format in the `res_format` field.