WIP - Show filter topics based on search results
If a user has entered a search term, the topic filter will only
include topics of datasets that are included in the search results.

Similar to the publisher filter, we can get this information by
including the use of faceting in the Solr query. We can specify
the `extras_theme-primary` fielt as a facet, and only return any
where the minimum count is 1.

We have additional complexity where the field does not return
the name that should be displayed in the UI, for example:
["environ",
 "govern",
 "governmentspend",
 "spend",
 "and",
 "busi",
 "businessandeconomi",
 "economi",
 "health"]

The list also includes duplication of names that are split by whitespace
e.g. the topic "Business and economy" is returned as "busi", "and", "economi"
as well as "businessandeconomi". We need to be able to map each topic to it's
display name. We could created another custom index (similar to site_id:dgu_organisations)
and the response from Solr would look something like:
docs:[
       {
         name: "govern",
         title: "Government",
       },
       {
         name: "governmentspend",
         title: "Government spending",
       },
       {
         name: "businessandeconomi",
         title: "Business and economy",
       },
     ]