WIP - Show filter formats based on search results
If a user has entered a search term or selected a filter option,
the format filter will only include formats for datafiles of
datasets that are included in the search results.

We can get this information by including the use of faceting in
the Solr query. We can specify the `res_format` field as a facet,
and only return any where the minimum count is 1.

The facet data returned includes a number of formats that aren't
valid (users are able to add this manually/part of a data inport).
We need to clean up this data so that it maps to valid formats,
and enforce some validation so that users can't add custom formats
in future.

Some example invalid formats:
"WWW:LINK-1.0-http--link", "WWW:DOWNLOAD-1.0-http--download",
"https://www.iana.org/assignments/media-types/application/json",
"Information and Download"