Add pagination controls to search results page
Solr caps the number of results returned to 10 by default, but we
want to display 20 per page. We can use Solr's `paginate` method,
which allows you to pass the page number and number of results
returned as parameters of the Solr query.

Kaminari gets confused on what the previous and next pages should
be (it's expecting certain variable names to be available but Solr
doesn't provide these). The RSolr docs recommend using the
`will_paginate` gem which fixes this issue.