Make topic filter only include topics of returned datasets
Topic filter should only include topics for datasets that are included in the
search results.

If a user has entered a search term and/or has selected some filters, we should
expect the topics filter to only include topics of datasets that are included
in the search results.

The facet count list duplication of names that are split by whitespace e.g. the
topic Business and economy is returned as busi, and, economi as well as
businessandeconomi. We decided to map each topic to it's display name.  There's
an issue with the and which could be for either "Crime and justice" or "Towns
and cities" or "Business and economy".

The root cause of this behaviour is the fact that extras_theme-primary appears
to be a tokenized field (text-type field) and the query is returning tokens as
facets. See:
https://github.com/hectorcorrea/solr-for-newbies/blob/main/tutorial.md#facets.
We decided not to attempt changing the field type to a string.

There's an issue with the `and` token which could be for either "Crime and
justice" or "Towns and cities" or "Business and economy". As a workaround based
on the facets counts we can actually discard and token because the facet counts
are the same for each group:

        "and",5908,
        "environ",5704,
        "citi",4370,
        "town",4370,
        "townsandc",4370,
        "govern",4176,
        "societi",2240,
        "map",2118,
        "health",1823,
        "governmentspend",1822,
        "spend",1822,
        "educ",1025,
        "busi",917,
        "businessandeconomi",917,
        "economi",917,
        "transport",762,
        "crime",621,
        "crimeandjustic",621,
        "justic",621,
        "defenc",181,
        "digit",16,
        "digitalservicesperform",16,
        "perform",16,
        "servic",16,
        "data",3,
        "governmentreferencedata",3,
        "refer",3,

`city` `town` `townsandc` all 4370
`crime` `crimeandjustic` `justic` all 621
`busi` `businessandeconomi` `economi` all 917

The count for and equals the sum of above: 4370 + 917 + 621 = 5908