Refactor SolrSearchController to improve performance
The before action used in this way is considered an anti-pattern as it’s not
necessary here.

The refactored solr_search_response method uses memoization which ensures that
the Search::Solr.search(params) call only runs once per request and caches the
result in @solr_search_response.  This approach improves efficiency as it is
called multiple times within the action.