Make format filter only include main formats of returned datasets
Format filter should only include valid formats for datafiles of datasets that
are included in the search results.

If a user has entered a search term and/or has selected some filters, we should
expect the formats filter to only include formats of datafiles of datasets that
are included in the search results.

We can get this information by including the use of faceting in the Solr query.
We can specify the res_format field as a facet, and only return any where the
minimum count is 1.

The facet data returned includes a number of formats that aren't valid or
display friendly (users are able to add this manually/part of a data import).
We need to clean up this data so that it maps to valid formats.

All available formats and a count Solr UI -
/solr/#/ckan/query?q=*:*&q.op=OR&indent=true&facet=true&facet.field=res_format&facet.sort=count&facet.mincount=1&fl=res_format&rows=1