Ensure OGL Licence filter returns all relevant results
We used `uk-ogl` id to filter datasets that use the standard OGL licence
(currently version 3). Datasets with `OGL-UK-3.0` and `ogl` should also be
returned.

Using the wildcard search for the version number will make it easier
to maintain to eliminate the need to update the code when a new version is
released.
https://solr.apache.org/guide/8_7/the-standard-query-parser.html#wildcard-searches

Initially wanted to use wildcard search "license_id:*OGL*"
However "OGL-UK-3.0" are not returned. Example query with 2 such licences:
?q=&filters%5Bpublisher%5D=City+of+York+Council&filters%5Bformat%5D=ZIP&filters%5Blicence_code%5D=uk-ogl&sort=best

Fuzzy searches have max distance parameter of 2 so are not suitable either.

Ideally, we would want this query string parameter to be a boolean but it would
be a breaking change so keeping licence_id with the value of "uk-ogl”.

license_id facet count at the time or writing
      "license_id":[
        "uk-ogl",25775,
        "cc-by",975,
        "OGL-UK-3.0",402,
        "notspecified",119,
        "other-nc",49,
        "cc-zero",17,
        "other-closed",17,
        "other-pd",13,
        "unpublished",11,
        "cc-by-sa",6,
        "other-open",6,
        "odc-odbl",5,
        "odc-by",4,
        "odc-pddl",3,
        "__other__",2,
        "cc-nc",1,
        "ogl",1,
        "other-at",1]

Example Solr query to return datasets with specific licence ID:
/solr/ckan/select?fq=license_id%3A"OGL-UK-3.0"&indent=true&q.op=OR&q=*%3A*&rows=100