Search separately for each individual whitespace-separated term
Because we previously had the entire parameter in quotes Solr treated it as a
single term.

Now when
- `animal health` is entered, we are splitting it by whitespace and Solr should
  be searching for animal or health, and returning all results for both terms.
- "animal health" is entered, we should not be splitting by whitespaces

Parentheses were added as the field is only valid for the term that it directly
precedes, so the query `title:Do it right` will find only "Do" in the title
field. It will find "it" and "right" in the default field.

Unsure what the default field is for us. It should be defined as
`<defaultSearchField>FieldName</defaultSearchField>` in the solrconfig.xml:
https://github.com/alphagov/ckanext-datagovuk/blob/main/docker/solr/solrconfig.xml.

To apply two or more operators to a single field in a search, we need to group
the clauses within parentheses.
See:
- https://solr.apache.org/guide/6_6/the-standard-query-parser.html#TheStandardQueryParser-SpecifyingFieldsinaQuerytotheStandardQueryParser
- https://solr.apache.org/guide/6_6/the-standard-query-parser.html#TheStandardQueryParser-GroupingClauseswithinaField