Explicitly set Solr search operator to "OR"
For the search to work as expected the `q.op` parameter, which specifies the
default operator for query expressions has to be set as "OR". Setting this
parameter allows to override the default operator specified in the Schema.
Possible values are "AND" or "OR".

The default operator specified in the Schema is "OR" but we are setting it here
explicitly to avoid unforeseen consequences if this is changed in the schema.