Add handling for Solr 400s errors
When Solr returns one of the 400 errors[^1], we want the users to see a No
results page, as it happens for the OpenSearch version. An example scenario is
when a user searches for one of Solr special characters (as we are not escaping
them).

The filters will be empty in this case.

[^1]: https://github.com/rsolr/rsolr/blob/1651450/lib/rsolr/error.rb#L109-L130