Use uuid consistently for solr Dataset
We’re using uuid and id inconsistently. For OpenSearch Dataset model had both id and uuid.
But `id` was only used for related datasets but we decided to remove this feature.

The format corresponds to uuid and methods that use it as a parameter have “by_uuid” in the name. It will also be easier to switch to Solr if this attribute is kept the same.