Remove unused map previews code
This code was not used.  "We decided to stop providing map previews as there
was a need to do some extra dev work to support a change in the API which we
didn't have capacity for so we decided to just disable it"

This feature will not be ported to Solr DB as only essential maintenance is
performed on data.gov.uk.