Explicitly set Solr query operator to OR
This was commented out for testing and merged by mistake.  OR is the default
value so it's not going to change the behaviour of the search.

We don't have solrconfig.xml defined. We're using one defined on the image:
https://github.com/alphagov/ckanext-datagovuk/blob/main/docker/solr/2.10.Dockerfile#L1
We want to define in the query in case there's an unexpected change in the
config.