Ensure datasets render correctly if values are missing
ActionView::Template::Error
wrong argument type nil (expected String) (ActionView::Template::Error)
was raised when some values are missing in the Solr database.

We've seen this happen for 'summary'. In OpenSearch this value was coming in as
"No description provided" so no additional handling was required.
This method is used for both summary and custom license so we want to keep the
string generic.

https://govuk.sentry.io/issues/6188276899/?project=146189