Increase the limit on organization facet
We were not returning all organisations associated with returned datasets in
the Publisher filter because the upper limit on organization facet is too low.
It defaults to 100.
https://solr.apache.org/guide/6_6/faceting.html#Faceting-Thefacet.limitParameter

"There are currently around 1500 organisations and the query is currently set
to return a max of 1600. This feels reasonable as new organisations are not
added often." was mentioned in
https://github.com/alphagov/datagovuk_find/commit/76b7204cd0c086241ed21589dd728f7c7b4e1aae
as a justification for the number.