Handle missing organisations
A number of organisations is missing from the Solr index. We want to be notified about those so we can investigate the issue with the Solr index but we still want to render the dataset pages correctly.

The contact details and FOI information on those pages may not be showing until the the organisation is in the index.

This will resolve the following Sentry errors:

NoMethodError
undefined method `[]' for nil:NilClass (NoMethodError)

    @title = hash["title"]
                 ^^^^^^^^^

https://govuk.sentry.io/issues/6205677017/?project=1461892
https://govuk.sentry.io/issues/6194439420/?project=1461892
https://govuk.sentry.io/issues/6205675340/?project=1461892
https://govuk.sentry.io/issues/6205675293/?project=1461892